Colors Update 1
the game is now able to read colors, as long as there are 2 maps in the Maps folder (first one for logical stuff like blocks that stop you and second one is visual, for stuff that do nothing like background colors).

The AdvancedDraw method still doesn't support it and needs to be rehauled so when the player moves over a pained block, the block will keep its color.